TC: start tending towards meaning.
